# AtlasRoutes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AtlasRoutes API",
        "description": "API for managing tourist itineraries, destinations, and user favorites.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/itineraries": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "List itineraries",
                "operationId": "86675d12475f84f785e7563fa8b8a0d9",
                "responses": {
                    "200": {
                        "description": "Itineraries list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Itinerary"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Create a new itinerary",
                "operationId": "9ddf03b1ccb2fb24c86fa810f0614b38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItineraryCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itinerary created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary created successfuly"
                                        },
                                        "itinerary": {
                                            "$ref": "#/components/schemas/Itinerary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/{itinerary}": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Get itinerary details",
                "operationId": "31c9a4a6aed036f4c01bc094ca82f3af",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinerary details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Itinerary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Delete an itinerary",
                "operationId": "8e7fcfd3c35013ae4fed77c37b2d2be2",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinerary deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Update an itinerary",
                "operationId": "403a6313254eec39e64f7a5da26a0308",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItineraryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Itinerary updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Itinerary updated successfully"
                                        },
                                        "itinerary": {
                                            "$ref": "#/components/schemas/Itinerary"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/search": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Search itineraries by title keyword",
                "operationId": "cee90f84da3ed505d1c8bda6dbcc851f",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matching itineraries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Itinerary"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/itineraries/filter": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Filter itineraries by category and duration",
                "operationId": "ca1294414e96c42ce440cbb6edce8534",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "duration",
                        "in": "query",
                        "description": "Max duration in days",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered itineraries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Itinerary"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/itineraries/popular": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Get most popular itineraries",
                "operationId": "a718dfc1cb4768f1906d8ecf6009204f",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Max results (1-50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Popular itineraries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Itinerary"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "destination_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "name": {
                        "type": "string",
                        "example": "Visit the old medina"
                    },
                    "type": {
                        "type": "string",
                        "example": "place",
                        "enum": [
                            "place",
                            "activity",
                            "dish"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Destination": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 4
                    },
                    "itinerary_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Marrakesh"
                    },
                    "accommodation": {
                        "type": "string",
                        "example": "Riad Atlas",
                        "nullable": true
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Itinerary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Atlas Route Highlights"
                    },
                    "category": {
                        "type": "string",
                        "example": "adventure"
                    },
                    "duration": {
                        "type": "integer",
                        "example": 5
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/atlas.jpg",
                        "nullable": true
                    },
                    "favorites_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "destinations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Destination"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2026-03-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "ItineraryCreateRequest": {
                "required": [
                    "title",
                    "category",
                    "duration",
                    "destinations"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Atlas Route Highlights"
                    },
                    "category": {
                        "type": "string",
                        "example": "adventure"
                    },
                    "duration": {
                        "type": "integer",
                        "example": 5
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/atlas.jpg",
                        "nullable": true
                    },
                    "destinations": {
                        "type": "array",
                        "items": {
                            "required": [
                                "name",
                                "activities"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Marrakesh"
                                },
                                "accommodation": {
                                    "type": "string",
                                    "example": "Riad Atlas",
                                    "nullable": true
                                },
                                "activities": {
                                    "type": "array",
                                    "items": {
                                        "required": [
                                            "name",
                                            "type"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Visit the old medina"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "place",
                                                "enum": [
                                                    "place",
                                                    "activity",
                                                    "dish"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ItineraryUpdateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Atlas Route"
                    },
                    "category": {
                        "type": "string",
                        "example": "culture"
                    },
                    "duration": {
                        "type": "integer",
                        "example": 7
                    },
                    "image": {
                        "type": "string",
                        "example": "https://example.com/atlas.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}